Refactor trace archival helper utilities (#22934)
